Fix Spinner Timing (#258)
* fix

* rm conditional rendering

* anoter?

* program details manages its own spinner

* revert

* refactor

* refactor

---------
