+++ src/features/multiDatasetDownload/components/MultiDatasetDownloadContainer.js
@@ -6,6 +6,7 @@ import DefaultSubsetControlsLayout from '../../../shared/filtering/components/De
 import useSubsetFiltering from '../../../shared/filtering/hooks/useSubsetFiltering';
 import useMultiDatasetDownloadStore from '../stores/multiDatasetDownloadStore';
 import useRowCountStore from '../stores/useRowCountStore';
+import { aggregateDatasetMetadata } from '../utils/aggregateDatasetMetadata';
 import MultiDatasetDownloadTable from './MultiDatasetDownloadTable';
 import DownloadButton from './DownloadButton';
 import RowCountTotal from './RowCountTotal';
@@ -42,48 +43,11 @@ import { SpinnerWrapper } from '../../../Components/UI/Spinner';
  * The component automatically computes aggregate bounds across all datasets for filtering
  * and initializes internal Zustand stores for state management.
  */
-const MultiDatasetDownloadContainerInner = ({ datasetsMetadata }) => {
+const MultiDatasetDownloadContainerInner = ({ aggregateDatasetMetadata }) => {
   const { resetStore, getSelectedIds } = useMultiDatasetDownloadStore();
   const { fetchRowCountsForSelected, cancelPendingRequests } =
     useRowCountStore();
   const filteredItems = useFilteredItems();
-
-  // Compute aggregate dataset bounds for multi-dataset filtering
-  const aggregateDatasetMetadata = useMemo(() => {
-    if (!datasetsMetadata || datasetsMetadata.length === 0) {
-      return null;
-    }
-
-    const validDatasets = datasetsMetadata.filter(
-      (d) =>
-        d.Lat_Min !== undefined &&
-        d.Lat_Max !== undefined &&
-        d.Lon_Min !== undefined &&
-        d.Lon_Max !== undefined,
-    );
-
-    if (validDatasets.length === 0) return null;
-
-    return {
-      Lat_Min: Math.min(...validDatasets.map((d) => d.Lat_Min)),
-      Lat_Max: Math.max(...validDatasets.map((d) => d.Lat_Max)),
-      Lon_Min: Math.min(...validDatasets.map((d) => d.Lon_Min)),
-      Lon_Max: Math.max(...validDatasets.map((d) => d.Lon_Max)),
-      Depth_Min: Math.min(...validDatasets.map((d) => d.Depth_Min || 0)),
-      Depth_Max: Math.max(...validDatasets.map((d) => d.Depth_Max || 0)),
-      Time_Min: validDatasets.reduce(
-        (min, d) =>
-          !min || (d.Time_Min && d.Time_Min < min) ? d.Time_Min : min,
-        null,
-      ),
-      Time_Max: validDatasets.reduce(
-        (max, d) =>
-          !max || (d.Time_Max && d.Time_Max > max) ? d.Time_Max : max,
-        null,
-      ),
-      Temporal_Resolution: validDatasets[0]?.Temporal_Resolution || 'daily',
-    };
-  }, [datasetsMetadata]);
 
   // State for toggle controls (required by layout components)
   const [optionsState, setOptionsState] = useState({
@@ -189,6 +153,11 @@ const MultiDatasetDownloadContainer = ({ datasetShortNames }) => {
   const { datasetsMetadata, fetchDatasetsMetadata, isLoading } =
     useMultiDatasetDownloadStore();
 
+  // Compute aggregate dataset bounds for multi-dataset filtering
+  const aggregateMetadata = useMemo(() => {
+    return aggregateDatasetMetadata(datasetsMetadata);
+  }, [datasetsMetadata]);
+
   useEffect(() => {
     if (datasetShortNames && datasetShortNames.length > 0) {
       fetchDatasetsMetadata(datasetShortNames);
@@ -196,12 +165,18 @@ const MultiDatasetDownloadContainer = ({ datasetShortNames }) => {
   }, [datasetShortNames, fetchDatasetsMetadata]);
 
   if (isLoading || !datasetsMetadata || datasetsMetadata.length === 0) {
-    return <SpinnerWrapper message="Loading datasets..." />;
+    return <SpinnerWrapper />;
+  }
+
+  if (!aggregateMetadata) {
+    return <SpinnerWrapper />;
   }
 
   return (
     <SearchProvider items={datasetsMetadata} searchKeys={['Dataset_Name']}>
-      <MultiDatasetDownloadContainerInner datasetsMetadata={datasetsMetadata} />
+      <MultiDatasetDownloadContainerInner
+        aggregateDatasetMetadata={aggregateMetadata}
+      />
     </SearchProvider>
   );
 };
